+++ reticulator/changelog.md
@@ -85,3 +85,6 @@
 # 0.1.2-beta
  - Introduced a new @SingleResourceDefinition decorator
  - Fixed intellisense for @identifier, @format_version, and @ClassProperty
+ - Child Getters now return 'None' instead of throwing an error
+ - Streamlined the 'Triple' system, and removed 'exists()', since you can just check for None
+ - Switched all 'adders' to use kwargs